set default tmp path before test starts
