+++ tests/integration/cli/classification/test_classification.py
@@ -5,8 +5,6 @@
 
 import os
 from functools import wraps
-from pathlib import Path
-from tempfile import TemporaryDirectory
 
 import pytest
 import torch
@@ -74,12 +72,6 @@ args = {
 }
 
 otx_dir = os.getcwd()
-
-
-@pytest.fixture(scope="session")
-def tmp_dir_path():
-    with TemporaryDirectory() as tmp_dir:
-        yield Path(tmp_dir)
 
 
 MULTI_GPU_UNAVAILABLE = torch.cuda.device_count() <= 1
