+++ tests/integration/cli/segmentation/test_segmentation.py
@@ -6,8 +6,6 @@
 import os
 import shutil
 from functools import wraps
-from pathlib import Path
-from tempfile import TemporaryDirectory
 
 import pytest
 import torch
@@ -71,12 +69,6 @@ if TT_STABILITY_TESTS:
 else:
     templates = Registry("otx/algorithms/segmentation").filter(task_type="SEGMENTATION").templates
     templates_ids = [template.model_template_id for template in templates]
-
-
-@pytest.fixture(scope="session")
-def tmp_dir_path():
-    with TemporaryDirectory() as tmp_dir:
-        yield Path(tmp_dir)
 
 
 class TestToolsMPASemiSLSegmentation:
